requirements.txt
asgiref==3.5.0
Django==4.0.2
factory-boy==3.2.1
Faker==12.1.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
